refactoring describe to describe_functions
